requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
langchain
PyPDF2
